added timeout to notification story
